+++ packages/common-ethereum/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Moved cli method `getAbiInterface` to match with `EthereumNetworkModule` (#323)
+
 ## [4.0.0] - 2024-07-03
 ### Added
 - Add alias `parseProjectManifest`, also follow type of `INetworkCommonModule` (#318)