+++ packages/types/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add missing method `getAbiInterface` in type `EthereumNetworkModule` (#323)
+
 ## [3.8.0] - 2024-07-03
 ### Added
 - Add type `EthereumNetworkModule` to support eth module (#318)